# microservices/mano-dns/mano-dns-pdns-aggregator/mano-pdns-api/src/main/java/com/ubiqube/mano/pdns/model/Server.java
/*
 * PowerDNS Authoritative HTTP API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.15
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

package com.ubiqube.mano.pdns.model;

import java.util.Objects;

import org.jspecify.annotations.Nullable;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;

/**
 * Server
 */
@JsonPropertyOrder({
		Server.JSON_PROPERTY_TYPE,
		Server.JSON_PROPERTY_ID,
		Server.JSON_PROPERTY_DAEMON_TYPE,
		Server.JSON_PROPERTY_VERSION,
		Server.JSON_PROPERTY_URL,
		Server.JSON_PROPERTY_CONFIG_URL,
		Server.JSON_PROPERTY_ZONES_URL
})
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2025-05-23T10:34:53.990305454+02:00[Europe/Paris]", comments = "Generator version: 7.13.0")
public class Server {
	public static final String JSON_PROPERTY_TYPE = "type";
	@Nullable
	private String type;

	public static final String JSON_PROPERTY_ID = "id";
	@Nullable
	private String id;

	public static final String JSON_PROPERTY_DAEMON_TYPE = "daemon_type";
	@Nullable
	private String daemonType;

	public static final String JSON_PROPERTY_VERSION = "version";
	@Nullable
	private String version;

	public static final String JSON_PROPERTY_URL = "url";
	@Nullable
	private String url;

	public static final String JSON_PROPERTY_CONFIG_URL = "config_url";
	@Nullable
	private String configUrl;

	public static final String JSON_PROPERTY_ZONES_URL = "zones_url";
	@Nullable
	private String zonesUrl;

	public Server() {
	}

	public Server type(@Nullable final String type) {

		this.type = type;
		return this;
	}

	/**
	 * Set to “Server”
	 *
	 * @return type
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_TYPE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getType() {
		return type;
	}

	@JsonProperty(JSON_PROPERTY_TYPE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setType(@Nullable final String type) {
		this.type = type;
	}

	public Server id(@Nullable final String id) {

		this.id = id;
		return this;
	}

	/**
	 * The id of the server, “localhost”
	 *
	 * @return id
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_ID)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getId() {
		return id;
	}

	@JsonProperty(JSON_PROPERTY_ID)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setId(@Nullable final String id) {
		this.id = id;
	}

	public Server daemonType(@Nullable final String daemonType) {

		this.daemonType = daemonType;
		return this;
	}

	/**
	 * “recursor” for the PowerDNS Recursor and “authoritative” for the
	 * Authoritative Server
	 *
	 * @return daemonType
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_DAEMON_TYPE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getDaemonType() {
		return daemonType;
	}

	@JsonProperty(JSON_PROPERTY_DAEMON_TYPE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setDaemonType(@Nullable final String daemonType) {
		this.daemonType = daemonType;
	}

	public Server version(@Nullable final String version) {

		this.version = version;
		return this;
	}

	/**
	 * The version of the server software
	 *
	 * @return version
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_VERSION)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getVersion() {
		return version;
	}

	@JsonProperty(JSON_PROPERTY_VERSION)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setVersion(@Nullable final String version) {
		this.version = version;
	}

	public Server url(@Nullable final String url) {

		this.url = url;
		return this;
	}

	/**
	 * The API endpoint for this server
	 *
	 * @return url
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_URL)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getUrl() {
		return url;
	}

	@JsonProperty(JSON_PROPERTY_URL)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setUrl(@Nullable final String url) {
		this.url = url;
	}

	public Server configUrl(@Nullable final String configUrl) {

		this.configUrl = configUrl;
		return this;
	}

	/**
	 * The API endpoint for this server’s configuration
	 *
	 * @return configUrl
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_CONFIG_URL)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getConfigUrl() {
		return configUrl;
	}

	@JsonProperty(JSON_PROPERTY_CONFIG_URL)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setConfigUrl(@Nullable final String configUrl) {
		this.configUrl = configUrl;
	}

	public Server zonesUrl(@Nullable final String zonesUrl) {

		this.zonesUrl = zonesUrl;
		return this;
	}

	/**
	 * The API endpoint for this server’s zones
	 *
	 * @return zonesUrl
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_ZONES_URL)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getZonesUrl() {
		return zonesUrl;
	}

	@JsonProperty(JSON_PROPERTY_ZONES_URL)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setZonesUrl(@Nullable final String zonesUrl) {
		this.zonesUrl = zonesUrl;
	}

	@Override
	public boolean equals(final Object o) {
		if (this == o) {
			return true;
		}
		if ((o == null) || (getClass() != o.getClass())) {
			return false;
		}
		Server server = (Server) o;
		return Objects.equals(this.type, server.type) &&
				Objects.equals(this.id, server.id) &&
				Objects.equals(this.daemonType, server.daemonType) &&
				Objects.equals(this.version, server.version) &&
				Objects.equals(this.url, server.url) &&
				Objects.equals(this.configUrl, server.configUrl) &&
				Objects.equals(this.zonesUrl, server.zonesUrl);
	}

	@Override
	public int hashCode() {
		return Objects.hash(type, id, daemonType, version, url, configUrl, zonesUrl);
	}

	@Override
	public String toString() {
		StringBuilder sb = new StringBuilder();
		sb.append("class Server {\n");
		sb.append("    type: ").append(toIndentedString(type)).append("\n");
		sb.append("    id: ").append(toIndentedString(id)).append("\n");
		sb.append("    daemonType: ").append(toIndentedString(daemonType)).append("\n");
		sb.append("    version: ").append(toIndentedString(version)).append("\n");
		sb.append("    url: ").append(toIndentedString(url)).append("\n");
		sb.append("    configUrl: ").append(toIndentedString(configUrl)).append("\n");
		sb.append("    zonesUrl: ").append(toIndentedString(zonesUrl)).append("\n");
		sb.append("}");
		return sb.toString();
	}

	/**
	 * Convert the given object to string with each line indented by 4 spaces
	 * (except the first line).
	 */
	private String toIndentedString(final Object o) {
		if (o == null) {
			return "null";
		}
		return o.toString().replace("\n", "\n    ");
	}

}